Crate faer_entity

Re-exports

pub use pulp;

Modules

complex_split

Utilities for split complex number types whose real and imaginary parts are stored separately.

Structs

ComplexConjGroup

ComplexGroup

IdentityGroup

NoSimd

Symbolic

Traits

ComplexField

Unstable trait containing the operations that a number type needs to implement.

Conjugate

Trait for types that may be implicitly conjugated.

Entity

Unstable core trait for describing how a scalar value may be split up into individual component.

ForCopyType

ForDebugType

ForType

RealField

Unstable trait containing the operations that a real number type needs to implement.

SimdCtx

SimpleEntity

UniversalReborrow

UniversalReborrowMut

Functions

from_copy

into_copy

one_simd_as_slice

simd_as_slice

simd_as_slice_unit

simd_index_as_slice

slice_as_mut_simd

slice_as_simd

Type Aliases

GroupCopyFor

GroupDebugFor

GroupFor

IndexFor

SimdGroupFor

SimdIndexFor

SimdMaskFor

SimdUnitFor

UnitFor