Module faer_entity::complex_split

Utilities for split complex number types whose real and imaginary parts are stored separately.

Structs

ComplexConj

This structure contains the real and imaginary parts of an implicitly conjugated value.

ComplexConjIter

This structure contains a pair of iterators that allow simultaneous iteration over the real and imaginary parts of a collection of implicitly conjugated complex values.

ComplexIter

This structure contains a pair of iterators that allow simultaneous iteration over the real and imaginary parts of a collection of complex values.