Trait faer_entity::Conjugate

pub unsafe trait Conjugate: Entity {
    type Conj: Entity + Conjugate<Conj = Self, Canonical = Self::Canonical>;
    type Canonical: Entity + Conjugate<Canonical = Self::Canonical>;

    // Required method
    fn canonicalize(self) -> Self::Canonical;
}

Trait for types that may be implicitly conjugated.

Safety

The associated types and functions must fulfill their respective contracts.

Required Associated Types

type Conj: Entity + Conjugate<Conj = Self, Canonical = Self::Canonical>

Must have the same layout as Self, and Conj::Unit must have the same layout as Unit.

type Canonical: Entity + Conjugate<Canonical = Self::Canonical>

Must have the same layout as Self, and Canonical::Unit must have the same layout as Unit.

Required Methods

fn canonicalize(self) -> Self::Canonical

Performs the implicit conjugation operation on the given value, returning the canonical form.

Object Safety

This trait is not object safe.